Refactor namespace from AppSettings to Configuration
