+++ source/Databricks/source/Jobs.UnitTests/Configuration/DatabricksJobsOptionsTests.cs
@@ -12,15 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.AppSettings;
+using Energinet.DataHub.Core.Databricks.Jobs.Configuration;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
+using FluentAssertions;
+using Xunit;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.UnitTests.AppSettings;
+namespace Energinet.DataHub.Core.Databricks.Jobs.UnitTests.Configuration;
 
-public class OptionsTests
+public class DatabricksJobsOptionsTests
 {
     [Theory]
-    [InlineAutoMoqData(typeof(DatabricksSqlStatementOptions), 5, "WorkspaceUrl", "WorkspaceToken", "WarehouseId", "DatabricksHealthCheckStartHour", "DatabricksHealthCheckEndHour")]
+    [InlineAutoMoqData(typeof(DatabricksJobsOptions), 5, "WorkspaceUrl", "WorkspaceToken", "WarehouseId", "DatabricksHealthCheckStartHour", "DatabricksHealthCheckEndHour")]
     public void Options_HaveTheCorrectSettingNamesAndNumberOfSettings(Type sut, int settingsCount, params string[] expectedNames)
     {
         // Arrange & Act
