+++ source/Databricks/source/Jobs.UnitTests/Internal/JobsApiClientTests.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.Jobs.AppSettings;
+using Energinet.DataHub.Core.Databricks.Jobs.Configuration;
 using Energinet.DataHub.Core.Databricks.Jobs.Internal;
 using FluentAssertions;
 using Microsoft.Extensions.Options;
