+++ source/Databricks/source/SqlStatementExecution.IntegrationTests/Fixtures/DatabricksSchemaManager.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.AppSettings;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Configuration;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
