+++ source/Databricks/source/SqlStatementExecution/Diagnostics/HealthChecks/DatabricksSqlStatementApiHealthCheck.cs
@@ -17,7 +17,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.AppSettings;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Configuration;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Options;
 using NodaTime;
