+++ source/Databricks/source/SqlStatementExecution/Diagnostics/HealthChecks/DatabricksSqlStatementsApiHealthCheckBuilderExtensions.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.AppSettings;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using Microsoft.Extensions.Options;
